+++ RELEASES.md
@@ -1139,3 +1139,13 @@ Based on:
 - [typescript v0.6.0-alpha.21] .
 ### Releases
 - [NPM v0.6.0-alpha.21] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.21 - .
+
+## 2026-02-25 02:18:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.728.0 (2.839.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.0-alpha.22] .
+### Releases
+- [NPM v0.6.0-alpha.22] https://www.npmjs.com/package/cribl-control-plane/v/0.6.0-alpha.22 - .